==> 02-1_refine_stages (lab).ipynb <==
## Objectives
- 플레이 로그 테이블들 8개에 대해서 앞서 select된 sample player들의 로그만 선택한다.
- 이들을 모두 합쳐서 stages.csv 파일로 저장한다.
- characters_stages_n.csv 파일들이 기본적으로 용량이 커서 이 작업은 연구실 PC로 실시했다.

from config import *

players = pd.read_csv("../data/sample_players.csv", encoding="utf-8")

len(players)

* raw stage 파일들을 로컬 PC에서 data/stages 폴더 안에 넣어두었다.

for n in range(1,8):
    n_stage = str(n)
    stages = pd.read_csv("../data/stages/characters_stages_"+n_stage+".csv").drop(columns=["id"])
    stages = simple_preprocess(stages)
    print(n, "before:", len(stages))
    
    stages = stages[stages["player_slug"].isin(players["player_slug"].unique())]
    stages = stages.drop_duplicates().reset_index(drop=True)
    print(n, "after:", len(stages))
    
    stages.to_csv("../data/stages/characters_stages_"+n_stage+"_ref.csv", index=None, encoding="utf-8")

* Concat at once

stages = pd.concat(map(pd.read_csv, ['../data/stages/characters_stages_1_ref.csv','../data/stages/characters_stages_2_ref.csv',
                                     '../data/stages/characters_stages_3_ref.csv','../data/stages/characters_stages_4_ref.csv',
                                     '../data/stages/characters_stages_5_ref.csv','../data/stages/characters_stages_6_ref.csv',
                                     '../data/stages/characters_stages_7_ref.csv']))

len(stages)

stages = stages.drop_duplicates().sort_values(by="event_datetime")

len(stages)

stages.to_csv("../data/stages.csv", index=None, encoding="utf-8")
