==> 02-4_process_final_touch.ipynb <==
from config import *

levelups_ref = pd.read_csv("../data/levelups_panel_2.csv", encoding="utf-8")
# levelups_ref = pd.read_csv("../data/levelups_panel_3.csv", encoding="utf-8") # work-in-progress (main quest)
levelups_ref["event_datetime"] = pd.to_datetime(levelups_ref["event_datetime"])
levelups_ref = levelups_ref.sort_values(["player_slug","level_from"]).reset_index(drop=True)

levelups_ref["player_slug"].nunique()

# 일부 레벨업이 관찰되지 않는 플레이어가 샘플에서 제외되었는지 마지막 체크
a = levelups_ref.groupby(["player_slug","character_slug"])["level_from"].max().reset_index()
b = levelups_ref.groupby(["player_slug","character_slug"]).size().to_frame("n").reset_index()
l_error = pd.merge(a,b).query("level_from != n").player_slug.unique()
print(len(l_error))

levelups_ref = levelups_ref.query("player_slug not in @l_error").reset_index(drop=True)
levelups_ref["player_slug"].nunique()

* Total hours played

# levelups_ref["playtime_total"] = levelups_ref[["playtime_main","playtime_side"]].sum(axis=1)
# levelups_ref["playtime_total_cum"] = levelups_ref.groupby(["player_slug","character_slug"])["playtime_total"].cumsum() # cum

* Highest character level

tmp = levelups_ref.groupby(["player_slug","character_slug"])["level_to"].max().to_frame("highest_level").reset_index()
levelups_ref = pd.merge(tmp, levelups_ref, how="left")

* Drop out status (현재 플레이 중인 단계인지, 아니면 게임을 그만두었는지 status check)

# if last level up happened in a week, then playing
tmp = np.where(levelups_ref["level_to"]<levelups_ref["highest_level"], "complete",
               np.where((datetime.datetime.strptime(window_eday,"%Y-%m-%d") - levelups_ref["event_datetime"]).apply(lambda x: x.total_seconds()) < (3600*24*7), "playing",
                        "drop"))

pd.Series(tmp).value_counts()

tmp_ = levelups_ref[["player_slug","character_slug"]].copy()
tmp_["status"] = tmp
levelups_ref = pd.concat([tmp_, levelups_ref.drop(columns=["player_slug","character_slug"])], axis=1)

# switch highest_level and status position
tmp = levelups_ref["status"]
levelups_ref.drop(columns=["status"], inplace=True)
levelups_ref.insert(levelups_ref.columns.tolist().index("highest_level")+1, "status", tmp)

* Total purchases (Response variable)

levelups_ref

tmp = levelups_ref.groupby(["player_slug","character_slug"])[["purchase","ad_total"]].sum().reset_index()
tmp.columns = ["player_slug","character_slug","total_purchase","total_ad"]
tmp = pd.merge(levelups_ref[["player_slug","character_slug"]], tmp, how="left")

levelups_ref = pd.concat([tmp, levelups_ref.drop(columns=["player_slug","character_slug"])], axis=1)

* Export

levelups_ref

levelups_ref.to_csv("../data/levelups_panel_4.csv", index=None, encoding="utf-8")
