==> 03-1_process_first_session_level.ipynb <==
## Objectives
- Binger classification을 위한 준비 단계로서 필요한 노트북이다.
- 최초로 로그인 한 후 가장 첫 세션에서 도달한 레벨을 first_session_level로 정의한다.

from config import *

levelups_ref = pd.read_csv("../data/levelups_panel_4.csv", encoding="utf-8")

l_sample = levelups_ref.player_slug.unique()
len(l_sample)

* Login history

# login history: character
logins = pd.read_csv(data_dir+"characters_login.csv", encoding="utf-8")
logins.drop(columns=["id","player_id","character_id","event_timestamp"], inplace=True)
logins = simple_preprocess(logins)

logins = logins.query("player_slug in @l_sample").reset_index(drop=True)

# register date
tmp = levelups_ref.query("level_from==1")[["player_slug","character_slug","event_datetime_prev"]]
tmp = tmp.sort_values(by=["event_datetime_prev"]).reset_index(drop=True)
tmp.rename(columns={"event_datetime_prev":"register_datetime"}, inplace=True)
tmp["register_datetime"] = pd.to_datetime(tmp["register_datetime"])
tmp

# next login date

def match_next_login(row):
    logins_ = logins.query("player_slug==@row.player_slug & character_slug==@row.character_slug & event_datetime>@row.register_datetime")
    logins_ = logins_["event_datetime"].tolist()
    
    # if there is no next login
    if len(logins_)==0:
        return None
    
    # if there is next login(s)
    else:
        for login_ in logins_:
            dur = (login_ - row.register_datetime).total_seconds() / 3600
            if dur > 1: # 1시간 이내 재로그인은 무시
                break
            
        if dur <= 1:
            return None
        else:
            return login_

tmp["next_login_datetime"] = tmp.progress_apply(match_next_login, axis=1)

tmp = tmp.sort_values(by="next_login_datetime").reset_index(drop=True)
tmp

* Next login 있었던 경우

# last level up before the next login date
tmp_ = levelups_ref[["player_slug","character_slug","event_datetime","level_to"]].copy()
tmp_["event_datetime"] = pd.to_datetime(tmp_["event_datetime"])
tmp_ = tmp_.sort_values(by=["event_datetime"]).reset_index(drop=True)
tmp_.rename(columns={"event_datetime":"next_login_datetime"}, inplace=True)

tmp1 = pd.merge_asof(tmp.query("next_login_datetime==next_login_datetime"), tmp_,
                     on="next_login_datetime", by=["player_slug","character_slug"], direction="backward")
tmp1["level_to"].fillna(1, inplace=True)
tmp1

* Next login 없었던 경우

tmp2 = tmp.query("next_login_datetime!=next_login_datetime")
tmp2 = pd.merge(tmp2,
                levelups_ref[["player_slug","character_slug","highest_level"]].rename(columns={"highest_level":"level_to"}).drop_duplicates(),
                how="left")
tmp2

tmp = pd.concat([tmp1, tmp2]).sort_index()
tmp.rename(columns={"level_to":"first_session_level"}, inplace=True)
tmp["first_session_level"] = tmp["first_session_level"].astype(int)
tmp

levelups_ref_5 = pd.merge(tmp[["player_slug","character_slug","first_session_level"]], levelups_ref, how="right")
levelups_ref_5

levelups_ref_5.to_csv("../data/levelups_panel_5.csv", index=None, encoding="utf-8")
